+++ backend-shared/plannershared.h
@@ -2,7 +2,6 @@
 #ifndef PLANNERSHARED_H
 #define PLANNERSHARED_H
 #include <QObject>
-#include "core/pref.h"
 
 // This is a shared class (mobile/desktop), and contains the core of the diveplanner
 // without UI entanglement.
@@ -23,16 +22,6 @@ class plannerShared: public QObject {
 	Q_PROPERTY(int ascrate50 READ ascrate50 WRITE set_ascrate50 NOTIFY ascrate50Changed);
 	Q_PROPERTY(int ascrate75 READ ascrate75 WRITE set_ascrate75 NOTIFY ascrate75Changed);
 	Q_PROPERTY(int descrate READ descrate WRITE set_descrate NOTIFY descrateChanged);
-
-	// Planning data, no conversion but different origins
-	Q_PROPERTY(deco_mode planner_deco_mode READ planner_deco_mode WRITE set_planner_deco_mode NOTIFY planner_deco_modeChanged);
-	Q_PROPERTY(bool dobailout READ dobailout WRITE set_dobailout NOTIFY dobailoutChanged);
-	Q_PROPERTY(int reserve_gas READ reserve_gas WRITE set_reserve_gas NOTIFY reserve_gasChanged);
-	Q_PROPERTY(bool safetystop READ safetystop WRITE set_safetystop NOTIFY safetystopChanged);
-	Q_PROPERTY(int gflow READ gflow WRITE set_gflow NOTIFY gflowChanged);
-	Q_PROPERTY(int gfhigh READ gfhigh WRITE set_gfhigh NOTIFY gfhighChanged);
-	Q_PROPERTY(int vpmb_conservatism READ vpmb_conservatism WRITE set_vpmb_conservatism NOTIFY vpmb_conservatismChanged);
-
 public:
 	static plannerShared *instance();
 
@@ -43,15 +32,6 @@ public:
 	static int ascrate75();
 	static int descrate();
 
-	// Planning data, no conversion but different origins
-	static deco_mode planner_deco_mode();
-	static bool dobailout();
-	static int reserve_gas();
-	static bool safetystop();
-	static int gflow();
-	static int gfhigh();
-	static int vpmb_conservatism();
-
 public slots:
 	// Ascend/Descend data, converted to meter/feet depending on user selection
 	static void set_ascratelast6m(int value);
@@ -60,15 +40,6 @@ public slots:
 	static void set_ascrate75(int value);
 	static void set_descrate(int value);
 
-	// Planning data, no conversion but different origins
-	static void set_planner_deco_mode(deco_mode value);
-	static void set_dobailout(bool value);
-	static void set_reserve_gas(int value);
-	static void set_safetystop(bool value);
-	static void set_gflow(int value);
-	static void set_gfhigh(int value);
-	static void set_vpmb_conservatism(int value);
-
 signals:
 	// Ascend/Descend data, converted to meter/feet depending on user selection
 	void ascratelast6mChanged(int value);
@@ -76,15 +47,6 @@ signals:
 	void ascrate50Changed(int value);
 	void ascrate75Changed(int value);
 	void descrateChanged(int value);
-
-	// Planning data, no conversion but different origins
-	void planner_deco_modeChanged(deco_mode value);
-	void dobailoutChanged(bool value);
-	void reserve_gasChanged(int value);
-	void safetystopChanged(bool value);
-	void gflowChanged(int value);
-	void gfhighChanged(int value);
-	void vpmb_conservatismChanged(int value);
 
 private:
 	plannerShared() {}
